Support Emoji Keyboard
Emoji Keyboard is now supported in a very limited fashion.
There are only some 800 emoji's and the icon set is that of google.
The categories are somewhat messed up and need to be fixed.
The access to the emoji bord is beyond stupid. To access you need to
switch to the selection keyboard and then press the button labled 'S'.
Yeah I know it's stupid.