#24 Break main keyboard view into smaller sub components
In order to be able to move the sidebar, it needs to be a self defined component, so main keyboard view has been divided into 2 components.

Also, to make sure that the layout of the main keyboard make sense, no mater the position of the sidebar, tried to move the layout from linear to constraint.

With this commit, the code works but the main keyboard UI does not make sense visually.