Keyboard layout changes + removed some duplicate entries
The underscore and hyphen were listed in the file twice, so now they're not. I also fixed up the layout slightly.

I cheated and used my layout generator script to just recreate the file :).

Oh and sadly the "YO!! _\m/" was changed to "!!!" lol.

Also some things were in the wrong place, like some special characters like question mark was in between the "selection mode sequence" and the "space" button. They should just be with the other letters.

The letters are now all in the order you'd read them for each section, and this goes clockwise from the top right.

I've also put the Esperanto letters underneath the capital letters.

I've resubmitted the pull request because I'm stupid an uploaded the wrong file when I moved the Esperanto to be closer to the letters.