Set up Shared Preference Helper
In the last iteration, we had a lot of crashes because a single share preference variable type was changed.
This caused class cast exception when the new iteration was trying to load the existing value for the preference.
This could have been easily avoided with an exception check. Putting it in place.