🎨 chore: Add a button on the sidebar to access the clippad
The changes were made to add support for a new custom keycode, `SWITCH_TO_CLIPPAD_KEYBOARD`, which allows the user to switch to the clipboard keypad. The `CustomKeycode` enum was updated to include the new keycode. The `MainKeyboardView` class was modified to update the setup method for the clipboard button to handle the new keycode. Additionally, a new vector drawable for the clipboard icon was added, and the content description and image source for the clipboard button in the sidebar view were updated to reflect the change. The content description in the strings.xml file was also updated to match the new functionality.